feat(commands): Migrate home system commands to TypeScript
This commit continues the phased migration of the command system by converting the "Home System" batch to TypeScript.

The following command files have been migrated from JavaScript to TypeScript, adhering to the established patterns:
- `home.ts` (containing `/home`, `/homes`, `/delhome`, `/sethome`)
- `warp.ts` (containing `/warp`, `/addwarp`, `/delwarp`)

The command loader (`index.ts`) has been updated to reflect these changes. All new files have been successfully compiled and linted.
