+++ src/modules/commands/warp.ts
@@ -1,0 +1,160 @@
+import * as mc from '@minecraft/server';
+import { ActionFormData } from '@minecraft/server-ui';
+import { CustomCommand, CommandExecutor } from './commandManager.js';
+import { errorLog } from '../../core/logger.js';
+import * as warpsManager from '../../core/warpsManager.js';
+import { getConfig } from '../../core/configManager.js';
+import { startTeleportWarmup } from '../../core/utils.js';
+import { setCooldown } from '../../core/cooldownManager.js';
+import { sendMessage } from '../../core/messaging.js';
+import { constants } from '../../core/constants.js';
+
+const warpCommand: CustomCommand = {
+    name: 'warp',
+    description: 'Teleports you to a set warp location.',
+    aliases: ['warps'],
+    permissionLevel: 1024,
+    hasCooldown: true,
+    cooldownId: 'warp',
+    parameters: [
+        { name: 'warpName', type: 'string', optional: true }
+    ],
+    execute: (executor: CommandExecutor, args: Record<string, any>) => {
+        if (!(executor instanceof mc.Player)) {return;}
+        const config = getConfig();
+        if (!config.warps.enabled) {
+            sendMessage(constants.warpsDisabled, executor);
+            return;
+        }
+
+        const teleportToWarp = (warpName: string) => {
+            const warpLocation = warpsManager.getWarp(warpName);
+            if (!warpLocation) {
+                sendMessage(`§cWarp '${warpName}' not found.`, executor);
+                return;
+            }
+
+            const warmupSeconds = config.warps.teleportWarmupSeconds;
+            const teleportLogic = () => {
+                try {
+                    executor.teleport(warpLocation, { dimension: mc.world.getDimension(warpLocation.dimensionId) });
+                    sendMessage(`§aTeleported to warp '${warpName}'.`, executor);
+                    setCooldown(executor, 'warp');
+                } catch (e: any) {
+                    sendMessage(`§cFailed to teleport. Error: ${e.message}`, executor);
+                    errorLog(`[/warp] Failed to teleport: ${e.stack}`);
+                }
+            };
+            startTeleportWarmup(executor, warmupSeconds, teleportLogic, `warp '${warpName}'`);
+        };
+
+        const warpNameArg = args.warpName as string | undefined;
+        if (warpNameArg) {
+            teleportToWarp(warpNameArg);
+            return;
+        }
+
+        const warpList = warpsManager.listWarps();
+
+        if (warpList.length === 0) {
+            sendMessage('§cThere are no warps set.', executor);
+            return;
+        }
+
+        const form = new ActionFormData()
+            .title('Teleport to a Warp')
+            .body('Select a warp to teleport to:');
+
+        warpList.forEach((warpName: string) => {
+            const location = warpsManager.getWarp(warpName);
+            if (location) {
+                form.button(`${warpName}\n§7(X: ${location.x.toFixed(2)}, Y: ${location.y.toFixed(2)}, Z: ${location.z.toFixed(2)})`);
+            }
+        });
+
+        form.show(executor).then(response => {
+            if (response.canceled) {return;}
+            if (response.selection === undefined) {return;}
+            const selectedWarp = warpList[response.selection];
+            teleportToWarp(selectedWarp);
+        }).catch(e => errorLog(`[/warp UI] ${e.stack}`));
+    }
+};
+
+const addWarpCommand: CustomCommand = {
+    name: 'addwarp',
+    description: 'Creates a new warp at your current location or at specified coordinates.',
+    aliases: ['setwarp'],
+    permissionLevel: 1, // Admin
+    parameters: [
+        { name: 'warpName', type: 'string' },
+        { name: 'x', type: 'int', optional: true },
+        { name: 'y', type: 'int', optional: true },
+        { name: 'z', type: 'int', optional: true }
+    ],
+    execute: (executor: CommandExecutor, args: Record<string, any>) => {
+        if (!(executor instanceof mc.Player)) {return;}
+        const { warpName, x, y, z } = args as { warpName: string, x?: number, y?: number, z?: number };
+        const hasX = x !== undefined && x !== null;
+        const hasY = y !== undefined && y !== null;
+        const hasZ = z !== undefined && z !== null;
+
+        let location;
+
+        if (hasX && hasY && hasZ) {
+            location = { x, y, z };
+        } else if (!hasX && !hasY && !hasZ) {
+            location = executor.location;
+        } else {
+            sendMessage('§cYou must provide all three coordinates (x, y, z) or none to use your current location.', executor);
+            return;
+        }
+
+        const result = warpsManager.setWarp(warpName, location, executor.dimension.id);
+        sendMessage(result.success ? `§a${result.message}` : `§c${result.message}`, executor);
+    }
+};
+
+const delWarpCommand: CustomCommand = {
+    name: 'delwarp',
+    description: 'Deletes an existing warp.',
+    permissionLevel: 1, // Admin
+    parameters: [
+        { name: 'warpName', type: 'string', optional: true }
+    ],
+    execute: (executor: CommandExecutor, args: Record<string, any>) => {
+        if (!(executor instanceof mc.Player)) {return;}
+
+        const deleteWarpByName = (warpName: string) => {
+            const result = warpsManager.deleteWarp(warpName);
+            sendMessage(result.success ? `§a${result.message}` : `§c${result.message}`, executor);
+        };
+
+        const warpNameArg = args.warpName as string | undefined;
+        if (warpNameArg) {
+            deleteWarpByName(warpNameArg);
+            return;
+        }
+
+        const warpList = warpsManager.listWarps();
+        if (warpList.length === 0) {
+            sendMessage('§cThere are no warps to delete.', executor);
+            return;
+        }
+
+        const form = new ActionFormData()
+            .title('Delete a Warp')
+            .body('Select a warp to delete:');
+
+        warpList.forEach((warpName: string) => form.button(warpName));
+
+        form.show(executor).then(response => {
+            if (response.canceled) {return;}
+            if (response.selection === undefined) {return;}
+            const selectedWarp = warpList[response.selection];
+            deleteWarpByName(selectedWarp);
+        }).catch(e => errorLog(`[/delwarp UI] ${e.stack}`));
+    }
+};
+
+export default [warpCommand, addWarpCommand, delWarpCommand];
